Playtest Recent Replays
Added a new readout to the Playtesting tab that shows the last 100 playtest games.
Overhauled the game history database access to utilize a POJO reader (sql2o) and a json serializer (gson).  The serialized output is piped to the client as json and converted to a table using a template library (json2html).  This pipeline is more or less what future client-server DB operations should look like, as it takes far less code to write and maintain.