Added Discord integration
Overhauled the DOM for chat so that the actual definitions are in html and css, and javascript is reduced to only functionality (where possible).
Added a button that switches between Discord and Legacy chat.
Added an API for retrieving player info (to set the username properly in Discord)

(cherry picked from commit e3b790fa2b6f25267abd06d76083e9a691522d89)