Added target format to the deck definition
Overhauled all references to deck construction to include target format.
Altered deck builder to only show the validations for the selected target format.
Expanded the deck builder validation to show all problems with the deck instead of just one.
Added target format to the list of decks.
Expanded the error message in the game hall when an invalid deck is chosen for a given format.
Added the target format into the deck selection dropdown in the main hall.